fix: rename ModelCard in AiModelCard
