+++ wipp-backend-data/src/main/java/gov/nist/itl/ssd/wipp/backend/data/aimodelcard/tensorflow/ModelDetails.java
@@ -27,26 +27,37 @@
  * cause risk of injury or damage to property. The software developed by NIST
  * employees is not subject to copyright protection within the United States.
  */
-package gov.nist.itl.ssd.wipp.backend.data.modelcard;
+package gov.nist.itl.ssd.wipp.backend.data.aimodelcard.tensorflow;
 
-import gov.nist.itl.ssd.wipp.backend.core.model.auth.PrincipalFilteredRepository;
+import org.springframework.data.mongodb.core.mapping.Document;
 
-import io.swagger.v3.oas.annotations.tags.Tag;
-import org.springframework.data.mongodb.repository.Query;
-import org.springframework.data.rest.core.annotation.RepositoryRestResource;
-import org.springframework.data.rest.core.annotation.RestResource;
+@Document
+public class ModelDetails
+{
+    /***************** ATTRIBUTE(S) *****************/
+    private String name;
+    private String overview;
+    private Version version;
+    private Owners owners;
+    private String licence;
+    //private String reference;
+    private String citation;
 
-/**
- * @author Antoine Meyer <antoine.meyer at nist.gov>
- */
-@Tag(name="ModelCard Entity")
-@RepositoryRestResource
-public interface ModelCardRepository extends PrincipalFilteredRepository<ModelCard, String> {
+    /***************** CONSTRUCTOR(S) *****************/
+    public ModelDetails(String na, String ov, Version ve, Owners ow, String li, String ci)
+    {
+        this.name = na;
+        this.overview = ov;
+        this.version = ve;
+        this.owners = ow;
+        this.licence = li;
+        this.citation = ci;
+    }
 
-    @Override
-    @RestResource(exported = false)
-    void delete(ModelCard mc);
-
-    @Query("{aiModelId:'?0'}")
-    ModelCard findModelCardByAiModelId(String aiModelId);
+    /***************** METHOD(S) *****************/
+    public String getName() { return name; }
+    public String getOverview() { return overview; }
+    public Version getVersion() { return version; }
+    public Owners getOwners() { return owners; }
+    public String getCitation() { return citation; }
 }
